dodanie obslugi jwt i refresh tokena, dodanie logowania do frontendu

+++ climb2gether - backend/climb2gether - backend/climb2gether - backend/Controllers/V1/PostController.cs
@@ -6,11 +6,15 @@ using climb2gether___backend.Contracts;
 using climb2gether___backend.Contracts.V1.Requests;
 using climb2gether___backend.Contracts.V1.Responses;
 using climb2gether___backend.Domain;
+using climb2gether___backend.Extensions;
 using climb2gether___backend.Services;
+using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
-namespace climb2gether___backend.Controllers
+namespace climb2gether___backend.Controllers.V1
 {
+    [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
     public class PostController : Controller
     {
         private readonly IPostService _postService;
@@ -28,10 +32,16 @@ namespace climb2gether___backend.Controllers
         [HttpPut(ApiRoutes.Posts.Update)]
         public async Task<IActionResult> Update([FromRoute] Guid postId, [FromBody] UpdatePostRequest request)
         {
-            var post = new Post { 
-                Id = postId,
-                Name = request.Name
-            };
+            var userOwnsPost = await _postService.UserOwnsPost(postId, GetUserId());
+
+            if (!userOwnsPost)
+            {
+                return BadRequest(new { error = "You do not own this post" });
+            }
+
+            var post = await _postService.GetPostByIdAsync(postId);
+            post.Name = request.Name;
+
 
             var updated = await _postService.UpdatePostAsync(post);
             
@@ -39,6 +49,11 @@ namespace climb2gether___backend.Controllers
                 return Ok(post);
 
             return NotFound();
+        }
+
+        private string GetUserId()
+        {
+            throw new NotImplementedException();
         }
 
         [HttpGet(ApiRoutes.Posts.Get)]
@@ -56,6 +71,13 @@ namespace climb2gether___backend.Controllers
         [HttpDelete(ApiRoutes.Posts.Delete)]
         public async Task<IActionResult> Delete([FromRoute] Guid postId)
         {
+            var userOwnsPost = await _postService.UserOwnsPost(postId, GetUserId());
+
+            if (!userOwnsPost)
+            {
+                return BadRequest(new { error = "You do not own this post" });
+            }
+
             var deleted = await _postService.DeletePostAsync(postId);
 
             if(deleted)
@@ -67,7 +89,10 @@ namespace climb2gether___backend.Controllers
         [HttpPost(ApiRoutes.Posts.Create)]
         public async Task<IActionResult> Create([FromBody] CreatePostRequest postRequest)
         {
-            var post = new Post { Name = postRequest.Name };
+            var post = new Post {
+                Name = postRequest.Name,
+                UserId = HttpContext.GetUserId()
+            };
 
             await _postService.CreatePostAsync(post);
